fix: fixed code after mentor's review
