+++ package.json
@@ -1,7 +1,7 @@
 {
-    "name": "webpack-boilerplate",
+    "name": "covid-19-dashboard",
     "version": "1.0.1",
-    "description": "WebPack: HTML, CSS, SCSS, JS, Img, Fonts",
+    "description": "HTML, SCSS, JS, Img, Fonts",
     "main": "",
     "browserslist": [
         "> 0.25%, not dead"
