+++ src/components/global-table/utils/temp.js
@@ -3,5 +3,3 @@ function sortData(data, parameter) {
 }
 const test = [{ value: 1 }, { value: 2 }, { value: 5 }, { value: 3 }, { value: 4 }];
 const result = [{ value: 1 }, { value: 2 }, { value: 3 }, { value: 4 }, { value: 5 }];
-
-console.log(sortData(test, "value"));
