+++ src/components/graph-component/util/get-value.util.js
@@ -2,10 +2,10 @@ import { ONE_HUNDREED } from "../../../core/index";
 
 export function getValueCumulative(obj, population) {
     if (!population) {
-        return Object.values(obj || { error: "error" });
+        return Object.values(obj || {});
     }
     if (!obj) {
-        return { error: "error" };
+        return {};
     }
     return Object.values(obj)?.map((el) => {
         return (el / population) * ONE_HUNDREED;
@@ -13,17 +13,17 @@ export function getValueCumulative(obj, population) {
 }
 
 export function getKeys(obj, param) {
-    return Object.keys(obj?.[param] || { error: "error" });
+    return Object.keys(obj?.[param] || {});
 }
 
 export function getValueDays(obj, population) {
     if (!population) {
-        return Object.values(obj || { error: "error" })
+        return Object.values(obj || {})
             .map((item, index, arr) => item - (arr[index - 1] || 0));
     }
 
     if (!obj) {
-        return { error: "error" };
+        return {};
     }
 
     return Object.values(obj)
