+++ src/components/index.js
@@ -1,4 +1,3 @@
-export * from "./base-markup";
 export * from "./global-table/index";
 export * from "./graph-component/index";
 export * from "./country-list/index";
