+++ src/components/map/index.js
@@ -1,2 +1,2 @@
-export * from "./map.component";
+export * from "./map-covied.component";
 export * from "./legend/index";
