+++ src/components/util/index.js
@@ -3,3 +3,6 @@ export * from "./create-click-radio.util";
 export * from "./create-change-select-graph.util";
 export * from "./click-choose-country.util";
 export * from "./create-change-select-country.util";
+export * from "./sort-data.util"
+export * from "./update-data-for-hundred.utils"
+export * from "./get-global-division.util"
