+++ src/components/util/sort-data.util.js
@@ -1,0 +1,3 @@
+export function sortData(data, parameter) {
+    return data?.sort((a, b) => (a[parameter] > b[parameter] ? -1 : 1));
+}
